static metaprogramming to mediate AES classes
this fixes the api compatibility problem mentioned in last commit, and
tries to deal with the #ifdef introduced inside the driver
implementation by using the mediator in a more elegant way. allegedly.
to understand why include/aes.h is so convoluted, one needs to remember
that EPOS is compiled with GCC 4.4.4 and --std=c++0x, which is not
completely implemented [1]. you'll see why this is a big problem.

putting together the metaprogrammed conditional that acts as a parent
class selector was quite easy: just a matter of moving files around and
substituting AES_Common. however, to preserve compatibility through
constructors, one needs to explicitly inherit this method from the
parent class, so the cipher mode can be correctly configured.

it is quite easy to use alias declarations (`using parent::parent`) and
call it a day, but we need some way to refer to the parent class. a
templatized typedef could work, since the key size needs to be specified.
but that doesn't exist on GCC 4.4 yet, so a struct is created with just
a typedef typename (the latter because the type is dependent on a
template parameter), and this is apparently called a metafunction.

then we can use the aforementioned alias thing, right? not really, because
it's also not implemented. how about `child() : parent()`? that's called
constructor delegation and it's also not supported. inheriting the
constructor could also work, but it's, again, not supported. so we must
set the desired mode attribute directly. surely we could use an enum
forward declaration, since these are specific to each parent class?
the exercise of guessing if this is supported is left to the reader.

wondering why we can't just leave the #ifdef, we opt to create an enum
in the child class, with the constructor acting as a cast from this enum
to an integer, which is itself a public attribute in the parent class.

so much for object orientation and software engineering. also, the
padding problem still exists, and this was not tested on a real EPOS.

[1] https://gcc.gnu.org/gcc-4.4/cxx0x_status.html